remove cursed code
previously, the bar was responsible for telling cont/stop signal
listeners to halt. this is silly because if the stop one receives
stop, it should... stop. it should stop itself and the cont one.

now, these listeners are responsible for cleanly shutting down, so the
bar doesnt do any of that anymore and we get to remove the paragraph
comments about specificly timed signals breaking invariants and all
the other nonsense that was bc of that.

woohoo!