fix(main): ;-; rare time-specific bug
instead of sending confirmations and replies on the same notify, we
have one explicitly for halting and the other explicitly for
confirmation.

im not entirely sure why the bug was happening, but i think it has
something to do with the underlying implementation of Notify that i
missed because i didnt read the fine print.